Add parameters to fix the MIDAS API display
Add a set of parameters which will limit the fields that are shown
when a user accesses the listing of public users.  One side is the field
values that we want to capture information for, the other is to unset
the values that are not queried from the model before it is returned.